// src/interfaces/complaintEnums.ts
// Shared enums for complaint creation and update modes
export enum ComplaintCaptureMode {
    CREATE_BY_ACCOUNT = 'CREATE_BY_ACCOUNT',
    CREATE_BY_NEW_MOBILE = 'CREATE_BY_NEW_MOBILE',
    CREATE_BY_PARTY = 'CREATE_BY_PARTY',
    UPDATE = 'UPDATE',
    CREATE_BY_WIFI_TOURIST = 'CREATE_BY_WIFI_TOURIST',
    CREATE_BY_SUBREQUEST = 'CREATE_BY_SUBREQUEST',
    CREATE_BY_MWALLET_ACCOUNT = 'CREATE_BY_MWALLET_ACCOUNT',
    CREATE_BY_OTT = 'CREATE_BY_OTT',
    IAM_REQUEST = 'IAM_REQUEST'
}

export enum ComplaintUpdateMode {
    ACCOUNT = 'ACCOUNT',
    NEW_MOBILE = 'NEW_MOBILE',
    PARTY = 'PARTY',
    WIFITOURIST = 'WIFITOURIST',
    SUBREQUEST = 'SUBREQUEST',
    MWALLET = 'MWALLET'
}

export enum ProvisioningType{
    	BY_ACTIVE_ACCOUNT = 'BY_ACTIVE_ACCOUNT', 
        BY_INACTIVE_ACCOUNT = 'BY_INACTIVE_ACCOUNT', 
        BY_PARTY = 'BY_PARTY'
    }