// src/utils/complaints/enums/ActivityGroupCode_Enum.ts
export enum ActivityGroupCode {
    UNKOWN = 'UNKOWN',
    FCR = 'FCR',
    FCRADJUST = 'FCRADJUST',
    ROAMING = 'ROAMING',
    FEMTOCELL = 'FEMTOCELL',
    PROVISION = 'PROVISION',
    COVERAGE = 'COVERAGE',
    NEWSIM = 'NEWSIM',
    QOS = 'QOS',
    CUSTSERV = 'CUSTSERV',
    BILL = 'BILL',
    SUBSCRIPTION = 'SUBSCRIPTION',
    WIMAX = 'WIMAX',
    BITSTREAM = 'BITSTREAM',
    PROVISIONMANUAL = 'PROVISIONMANUAL',
    EPM = 'EPM',
    FRAUD = 'FRAUD',
    ADJUSTUCMS = 'ADJUSTUCMS',
    INDY = 'INDY',
    LOYALTY = 'LOYALTY',
    Request = 'Request',
    WIFITOURIST = 'WIFITOURIST',
    REQUEST_ACTIVE = 'REQUEST_ACTIVE',
    REQUEST_INACTIVE = 'REQUEST_INACTIVE',
    REQUEST_BOTH = 'REQUEST_BOTH',
    EWALLET_TECHNICAL = 'EWALLET_TECHNICAL',
    EWALLET_BILLING = 'EWALLET_BILLING',
    EWALLET_REQUEST = 'EWALLET_REQUEST',
    SWITCH_TV = 'SWITCH_TV',
    FOODORDERING = 'FOODORDERING',
    HLS = 'HLS',
    IAM_REQUEST = 'IAM_REQUEST',
    GOCHAT = 'GOCHAT',
    ELEARNING = 'ELEARNING',
    ELEARNING_REQUEST = 'ELEARNING_REQUEST',
    ELEARNING_COMPLAINT = 'ELEARNING_COMPLAINT',
    REQUEST_SPAM = 'REQUEST_SPAM'
}